requirements.txt
SpeechRecognition==3.8.1
Pillow==8.3.1
